use warnings to identify size_t -> int conversion and replace with size_t
+ fixing some implicit conversions to double with explicit casts

+++ headers/Class_Initial_Data.hpp
@@ -12,7 +12,7 @@
 class PressureVesselConfig {
 public:
 	bool IsSet;
-	int GasSpecies;
+	size_t GasSpecies;
 	double SampleSize;
 	double InitPressure;
 	double MaxPressure;
@@ -107,7 +107,7 @@ public:
 	/* Constant concentration option */
 	/* Single vector, to switch rates of specified species to zero */
 	bool ConstantConcentration;
-	vector< int > ConstantSpecies;
+	vector< size_t > ConstantSpecies;
 
 	// extra parameters for a TGA case
 	bool TGA_used;
