+++ source/mechanism_reduction/species_lumping/Average_Ea_k-fitted_Fast.cpp
@@ -11,12 +11,12 @@
 ReactionParameter Average_Ea_CalculateNewParametersFast(
 		vector< SingleReactionData >& Reactions,
 		double temperature,
-		int Reaction_Group_Size
+		size_t Reaction_Group_Size
 )
 {
 
 	ReactionParameter ParameterOutput;
-	int j;
+	size_t j;
 
 	vector <double> Group_k(3);
 	double total_Ea = 0;
