+++ source/mechanism_reduction/species_lumping/Average_Ea_k-fitted_Slow.cpp
@@ -11,7 +11,7 @@
 ReactionParameter Average_Ea_CalculateNewParametersSlow(
 		vector< SingleReactionData >& temp_reactions3,
 		double temperature,
-		int Reaction_Group_Size
+		size_t Reaction_Group_Size
 )
 {
 
@@ -22,7 +22,7 @@ ReactionParameter Average_Ea_CalculateNewParametersSlow(
 
 	//cout << "Slow Parameter Calculation Function \n";
 	ReactionParameter ParameterOutput;
-	int j;
+	size_t j;
 
 	vector <double> Group_k(41);
 	double fitted_Ea = 0;
@@ -31,8 +31,8 @@ ReactionParameter Average_Ea_CalculateNewParametersSlow(
 	for(j=0;j<Reaction_Group_Size;j++)
 	{
 		// work out k over all reactions
-		for(int k=0;k<41;k++){
-			double temp_mod = k - 20;
+		for(size_t k=0;k<41;k++){
+			double temp_mod = (double)k - 20.0;
 
 			Group_k[k] =
 					Group_k[k] +
@@ -54,8 +54,8 @@ ReactionParameter Average_Ea_CalculateNewParametersSlow(
 	vector <double> Ea_for_gradient(41);
 	//vector <double> lnT_for_gradient(41);
 
-	for(int k=0;k<41;k++){
-		double temp_mod = k - 20;
+	for(size_t k=0;k<41;k++){
+		double temp_mod = (double)k - 20.0;
 		// we do not need R as the internal Ea is in Kelvins
 		Ea_for_gradient[k] = -fitted_Ea/((temperature + temp_mod));
 		//lnT_for_gradient[k] = log(temperature + temp_mod);
@@ -70,7 +70,7 @@ ReactionParameter Average_Ea_CalculateNewParametersSlow(
 	//*
 	double gradient_k = 0;
 	double gradient_Ea = 0;
-	for(int k=1;k<41;k++){
+	for(size_t k=1;k<41;k++){
 		gradient_k = gradient_k +
 				(
 						(Group_k[k-1]-Group_k[k])
